% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getRange}
\alias{getRange}
\title{Identify range where IDs and marks are stored in HIS xlsx file}
\usage{
getRange(df, colIDs = "Matrikelnummer", colMarks = "Leistung")
}
\arguments{
\item{df}{data.frame containing the HIS template}

\item{colIDs}{character column name where IDS (Matrikelnummer) are stored in template}

\item{colMarks}{character column name where Marks (Noten) are stored in template}
}
\description{
Identify range where IDs and marks are stored in HIS xlsx file
}
