% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{marksFromHISxlsx}
\alias{marksFromHISxlsx}
\title{Extract Marks from HIS xlsx file}
\usage{
marksFromHISxlsx(df, ...)
}
\arguments{
\item{df}{data.frame containing the HIS template}

\item{...}{further arguements passed to `getRange`}
}
\value{
data.frame with columns `ID` and `mark`
}
\description{
Extract Marks from HIS xlsx file
}
