% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sourceMarksToHIS.R
\name{marksToHISxlsx}
\alias{marksToHISxlsx}
\title{Writes tabular ID-marks- data to HIS xlsx template}
\usage{
marksToHISxlsx(df, marksUpdated, ...)
}
\arguments{
\item{df}{data.frame containing the HIS template}

\item{marksUpdated}{data.frame tabular data with columns `ID` and `mark`}

\item{...}{further arguements passed to `getRange`}
}
\value{
updated HIS xlsx file
}
\description{
Writes tabular ID-marks- data to HIS xlsx template
}
\examples{
\dontrun{
# read xlsx file exported from HIS
require(openxlsx)
dfHIS <-  read.xlsx(xlsxFile = "HIS_xlsx_template.xlsx", colNames = FALSE,
                    skipEmptyRows = FALSE, skipEmptyCols = FALSE)
# read file with existing marks
# with colums 'ID' and 'mark'
dfMarks <- read.csv(file = "marks.csv")

# Workflow to update marks
marksHIS <- marksFromHISxlsx(dfHIS)
marksNew <- replaceMarks(marksHIS = marksHIS, dfMarks = dfMarks,
                         colID = "ID", colMarks = "mark")
dfHISnew <- marksToHISxlsx(df = dfHIS, marksUpdated = marksNew)

# write updated xlsx file to be imported to HIS
write.xlsx(dfHISnew, file = "HIS_updated.xlsx", colNames = FALSE, rowNames = FALSE)
}
}
