% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{replaceMarks}
\alias{replaceMarks}
\title{Replace Marks with existing ones based on ID mapping}
\usage{
replaceMarks(marksHIS, dfMarks, colID, colMarks, strNotAttended = "-")
}
\arguments{
\item{marksHIS}{data.frame marks extracted from HIS xlsx template}

\item{dfMarks}{data.frame with existing marks to be inserted into `marksHIS`}

\item{colID}{character column name containing the IDs in `marksHIS`}

\item{colMarks}{character column name containing the marks in `marksHIS`}

\item{strNotAttended}{character indicating HIS that student did not attend the exam}
}
\value{
updated marks (data.frame with columns `ID` and `mark`)
}
\description{
Replace Marks with existing ones based on ID mapping
}
\details{
Currently marks for `ID`s in `marksHIS` that aren't contained in `dfMarks`
are set to `strNotAttended` which is interpreted by HIS as *not attended*.
}
